Use UBI instead of UBI minimal as base image
So, when we used ubi minimal it looked like ld-musl-x86_64.so.1
interpreter was not present and becaucse of that even though if
the binary was present in the image we were  not able to execute.

Once we change UBI minimal to UBI and then install above mentioned
interpreter things seem to work.